backend/pyproject.toml
[project]
name = "order-builder"
version = "0.1.0"
description = ""
authors = [
    {name = "Vimal Saini",email = "vimal.saini25@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "marshmallow (>=3.26.1,<4.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "shopifyapi (>=12.7.0,<13.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pyotp (>=2.9.0,<3.0.0)",
    "boto3 (>=1.37.11,<2.0.0)",
    "python-magic (>=0.4.27,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.13"

